pyproject.toml
[project]
name = "eternal-backend"
version = "0.1.0"
dependencies = [
    "django",
    "django-cors-headers",
    "djangorestframework"
]
